added pid curve implementation added pid curve config refactored curve implementations into separate files
